Little fixes
